Train classifier only if not already trained

+++ train_classifier.py
@@ -1,3 +1,4 @@
+from os.path import isfile
 from sklearn import svm
 from sklearn.externals import joblib
 from tweet_sentiment import get_features
@@ -8,16 +9,17 @@ from tweet_listener import TweetListener
 import csv
 
 if __name__ == '__main__':
-    with open('data/dataset.csv') as dataset_file:
-        dataset = csv.reader(dataset_file)
-        features = []
-        classes = []
-        for tweet_data in dataset:
-            tweet = process_tweet(tweet_data[3])
-            tweet_features = get_features(tweet)
-            features.append(tweet_features)
-            classes.append(tweet_data[1])
+    if not isfile('data/classifier.pkl'):
+        with open('data/dataset.csv') as dataset_file:
+            dataset = csv.reader(dataset_file)
+            features = []
+            classes = []
+            for tweet_data in dataset:
+                tweet = process_tweet(tweet_data[3])
+                tweet_features = get_features(tweet)
+                features.append(tweet_features)
+                classes.append(tweet_data[1])
 
-    classifier = svm.SVC()
-    classifier.fit(features, classes)
-    joblib.dump(classifier, 'data/classifier.pkl')
+        classifier = svm.SVC()
+        classifier.fit(features, classes)
+        joblib.dump(classifier, 'data/classifier.pkl')
